requirements.txt
Django
djangorestframework
folium
phonenumbers
gunicorn
whitenoise
pymysql
django-cors-headers
djangorestframework-simplejwt
opencage
twilio